+++ packages/platforms/browser/tests/browser.test.ts
@@ -3,6 +3,8 @@
  * @jest-environment-options { "url": "https://bugsnag.com/browser-integration-tests" }
 */
 
+/* eslint-disable @typescript-eslint/no-var-requires */
+
 import { VALID_API_KEY } from '@bugsnag/js-performance-test-utilities'
 import type { BrowserConfiguration } from '../lib/config'
 import type { Client } from '@bugsnag/core-performance'
@@ -56,9 +58,7 @@ beforeEach(() => {
     mockFetch = createMockFetch()
     window.fetch = mockFetch
 
-    // eslint-disable-next-line @typescript-eslint/no-var-requires
     client = require('../lib/browser').default
-    // eslint-disable-next-line @typescript-eslint/no-var-requires
     bugsnag = require('@bugsnag/browser').default
   })
 
@@ -199,7 +199,7 @@ describe('Browser client integration tests', () => {
         autoInstrumentFullPageLoads: false,
         autoInstrumentNetworkRequests: false,
         autoInstrumentRouteChanges: false,
-        bugsnag
+        bugsnag: errorClient
       })
 
       await jest.runOnlyPendingTimersAsync()
@@ -247,7 +247,7 @@ describe('Browser client integration tests', () => {
         autoInstrumentFullPageLoads: false,
         autoInstrumentNetworkRequests: false,
         autoInstrumentRouteChanges: false,
-        bugsnag
+        bugsnag: errorClient
       })
 
       await jest.runOnlyPendingTimersAsync()
@@ -292,7 +292,7 @@ describe('Browser client integration tests', () => {
         autoInstrumentFullPageLoads: false,
         autoInstrumentNetworkRequests: false,
         autoInstrumentRouteChanges: false,
-        bugsnag
+        bugsnag: errorClient
       })
 
       await jest.runOnlyPendingTimersAsync()