+++ packages/platforms/react-native/lib/client.ts
@@ -51,7 +51,7 @@ const BugsnagPerformance = createClient({
   schema: createSchema(),
   spanAttributesSource,
   retryQueueFactory: createRetryQueueFactory(FileSystem),
-  platformExtensions
+  platformExtensions: (spanFactory, spanContextStorage) => platformExtensions(appStartTime, clock, spanFactory, spanContextStorage)
 })
 
 export default BugsnagPerformance