+++ test/react-native/features/fixtures/app/babel.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  presets: ['module:metro-react-native-babel-preset'],
-  plugins: ['@babel/plugin-transform-export-namespace-from', 'module:react-native-dotenv']
-}