+++ test/react-native/features/fixtures/scenario-launcher/scenarios/index.js
@@ -1,5 +1,7 @@
 export * as AppStartScenario from './AppStartScenario'
+export * as WithInstrumentedAppStartsScenario from './WithInstrumentedAppStartsScenario'
 export * as BackgroundSpanScenario from './BackgroundSpanScenario'
+export * as ErrorCorrelationScenario from './ErrorCorrelationScenario'
 export * as ManualSpanScenario from './ManualSpanScenario'
 export * as NavigationSpanScenario from './NavigationSpanScenario'
 export * as NestedSpansScenario from './NestedSpansScenario'