+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [v2.9.0] (2024-09-05)
+
+### Added
+
+- (react-native) Add new withInstrumentedAppStarts method to workaround automatically instrumented app start span issues [#497](https://github.com/bugsnag/bugsnag-js-performance/pull/497)
+
+### Changed
+
+- (core) Discard spans open for more than one hour [#494](https://github.com/bugsnag/bugsnag-js-performance/pull/494)
+- (core) use API key subdomain as default endpoint [#500](https://github.com/bugsnag/bugsnag-js-performance/pull/500)
+
 ## [v2.8.0] (2024-08-20)
 
 ### Added